feat(awino-ui): add dummy etherscan api key
The key is not a real key at the moment, because the implementation is not finished yet and the key is exposed insecurely in the frontend. Instead, an API endpoint must be implemented that performs the requests to Etherscan and reduces the risk that we run into rate limits with the etherscan api itself by means of rate limiting.